update keymap
There is a new layer to go to layer 1 from layer 0 - hold space to
go to layer 1, press space to get space.
Layer 1 no longer contains numbers. Instead numbers are replaced
with their shifted symbols.
When in layer 1 press and hold the space key to go to layer 3.
This mirrors the behaviour of the tilde key from the left half.